backend/requirements.txt
fastapi
uvicorn
python-multipart
prometheus_client
python-json-logger
PyJWT
openai
httpx
cryptography
pytest
requests
